+++ src/main/java/travelplanner/project/demo/global/security/CustomAuthenticationFilter.java
@@ -10,7 +10,7 @@ import org.springframework.security.authentication.UsernamePasswordAuthenticatio
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.AuthenticationException;
 import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
-import travelplanner.project.demo.member.Auth.LoginRequest;
+import travelplanner.project.demo.member.auth.LoginRequest;
 
 
 import java.io.IOException;
