--- src/main/java/travelplanner/project/demo/member/Auth/Role.java
@@ -1,5 +1,0 @@
-package travelplanner.project.demo.member.Auth;
-
-public enum Role {
-    MEMBER
-}
