+++ src/main/java/travelplanner/project/demo/member/auth/AuthController.java
@@ -1,4 +1,4 @@
-package travelplanner.project.demo.member.Auth;
+package travelplanner.project.demo.member.auth;
 
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
