+++ src/main/java/travelplanner/project/demo/member/auth/LoginRequest.java
@@ -1,4 +1,4 @@
-package travelplanner.project.demo.member.Auth;
+package travelplanner.project.demo.member.auth;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.AllArgsConstructor;
@@ -6,19 +6,16 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-@Schema(description = "회원가입 요청 DTO")
+@Schema(description = "로그인 요청 DTO")
 @Data
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class RegisterRequest {
-
-    @Schema(description = "닉네임", example = "유저1")
-    private String userNickname;
+public class LoginRequest {
 
     @Schema(description = "이메일", example = "user1@gmail.com")
     private String email;
 
-    @Schema(description = "비밀번호", example = "123456879")
+    @Schema(description = "비밀번호", example = "123456789")
     private String password;
 }
