+++ src/main/java/travelplanner/project/demo/member/auth/Role.java
@@ -1,0 +1,5 @@
+package travelplanner.project.demo.member.auth;
+
+public enum Role {
+    MEMBER
+}
